refactor: move ArrayHelper to Helpers/Array/ArrayHelper.php
